configuration scripts now in python & started fixing compiler errors msvc

+++ inc/sps/dataset.h
@@ -719,9 +719,9 @@ template <typename type_defs> class Dataset
         for( size_t uiI = 0; uiI < D; uiI++ )
             uiNumOverlaysTotal *= uiNumOverlays[ uiI ];
 
-        uint64_t uiNumOverlaySamples = std::max( 1ul, (uint64_t)std::log2( uiNumOverlaysTotal ) );
-        uint64_t uiNumPointSamples =
-            std::max( 1ul, 5 * (uint64_t)std::log2( xSortedPoints[ 0 ].uiEndIndex - xSortedPoints[ 0 ].uiStartIndex ) );
+        uint64_t uiNumOverlaySamples = std::max( (uint64_t)1, (uint64_t)std::log2( uiNumOverlaysTotal ) );
+        uint64_t uiNumPointSamples = std::max(
+            (uint64_t)1, 5 * (uint64_t)std::log2( xSortedPoints[ 0 ].uiEndIndex - xSortedPoints[ 0 ].uiStartIndex ) );
 
         std::tuple<uint64_t, uint64_t, uint64_t, uint64_t> tTotal{ };
         {
@@ -1146,14 +1146,14 @@ template <typename type_defs> class Dataset
             return std::numeric_limits<coordinate_t>::max( );
         else
             return std::min( xOverlays.vAxisSizes[ uiD ] - 1,
-                                    ( uiDataCoord - uiMinCoords[ uiD ] ) / uiSizeOverlays[ uiD ] );
+                             ( uiDataCoord - uiMinCoords[ uiD ] ) / uiSizeOverlays[ uiD ] );
     }
 
     pos_t overlayCoord( const pos_t& vPos ) const
     {
         pos_t vRet;
         for( size_t uiI = 0; uiI < D; uiI++ )
-            vRet[ uiI ] = overlayCoord(vPos[uiI], uiI);
+            vRet[ uiI ] = overlayCoord( vPos[ uiI ], uiI );
         return vRet;
     }
